+++ contracts/tool/src/index.ts
@@ -1,0 +1,3 @@
+export * from "./approval";
+export * from "./result";
+export * from "./spec";